requirements-dev.txt
Flask
tensorflow
Werkzeug
Pillow
PyYAML
